# Optimize performance of jaccard function