--- a/src/core_functions/scalar/string/jaccard.cpp
+++ b/src/core_functions/scalar/string/jaccard.cpp
@@ -2,7 +2,7 @@
 #include "duckdb/common/vector_operations/vector_operations.hpp"
 #include "duckdb/core_functions/scalar/string_functions.hpp"
 
-#include <array>
+#include <bitset>
 #include <ctype.h>
 
 namespace duckdb {
@@ -11,15 +11,14 @@ namespace {
 constexpr size_t MAX_SIZE = std::numeric_limits<unsigned char>::max() + 1;
 }
 
-static inline std::array<char, MAX_SIZE> GetSet(const string_t &str) {
-	std::array<char, MAX_SIZE> array_set;
-	array_set.fill(0);
+static inline std::bitset<MAX_SIZE> GetSet(const string_t &str) {
+	std::bitset<MAX_SIZE> array_set;
 
 	idx_t str_len = str.GetSize();
 	auto s = str.GetData();
 
 	for (idx_t pos = 0; pos < str_len; pos++) {
-		array_set[static_cast<unsigned char>(s[pos])] = 1;
+		array_set.set(static_cast<unsigned char>(s[pos]));
 	}
 	return array_set;
 }
@@ -28,17 +27,13 @@ static double JaccardSimilarity(const string_t &str, const string_t &txt) {
 	if (str.GetSize() < 1 || txt.GetSize() < 1) {
 		throw InvalidInputException("Jaccard Function: An argument too short!");
 	}
-	std::array<char, MAX_SIZE> m_str, m_txt;
+	std::bitset<MAX_SIZE> m_str, m_txt;
 
 	m_str = GetSet(str);
 	m_txt = GetSet(txt);
 
-	idx_t size_intersect = 0;
-	idx_t size_union = 0;
-	for (size_t i = 0; i < MAX_SIZE; ++i) {
-		size_intersect += m_str[i] & m_txt[i];
-		size_union += m_str[i] | m_txt[i];
-	}
+	idx_t size_intersect = (m_str & m_txt).count();
+	idx_t size_union = (m_str | m_txt).count();
 
 	return static_cast<double>(size_intersect) / static_cast<double>(size_union);
 }