Use `supplier.tbl` of ssb with column NAME and ADDRESS to test(2 million rows).

Before optimize:
```sql
D explain analyze select jaccard(column1, column2) from 'data.csv';

┌─────────────────────────────┐
│┌───────────────────────────┐│
│└───────────────────────────┘│
└─────────────────────────────┘
┌─────────────────────────────────────┐
│┌───────────────────────────────────┐│
││    Query Profiling Information    ││
│└───────────────────────────────────┘│
└─────────────────────────────────────┘
explain analyze select jaccard(column1, column2) from 'data.csv';
┌─────────────────────────────────────┐
│┌───────────────────────────────────┐│
││         Total Time: 0.687s        ││
│└───────────────────────────────────┘│
└─────────────────────────────────────┘
┌───────────────────────────┐
│      EXPLAIN_ANALYZE      │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│             0             │
│          (0.00s)          │
└─────────────_─────────────┘                              
┌─────────────┴─────────────┐
│         PROJECTION        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│ jaccard(column1, column2) │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (13.97s)         │
└─────────────_─────────────┘                              
┌─────────────┴─────────────┐
│READ_CSV_AUTO (MULTI-T...  │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          column1          │
│          column2          │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│        EC: 4916324        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (6.58s)          │
└───────────────────────────┘          
```

After optimize:
```sql
D explain analyze select jaccard(column1, column2) from 'data.csv';                                                                                                                                                 
                                                                                                                                                                                                                    
┌─────────────────────────────┐                                                                                                                                                                                     
│┌───────────────────────────┐│                                                                                                                                                                                     
│└───────────────────────────┘│                                                                                                                                                                                     
└─────────────────────────────┘                                                                                                                                                                                     
┌─────────────────────────────────────┐                                                                                                                                                                             
│┌───────────────────────────────────┐│                                                                                                                                                                             
││    Query Profiling Information    ││                                                                                                                                                                             
│└───────────────────────────────────┘│                                                                                                                                                                             
└─────────────────────────────────────┘                                                                                                                                                                             
explain analyze select jaccard(column1, column2) from 'data.csv';                                                                                                                                                   
┌─────────────────────────────────────┐                                                                                                                                                                             
│┌───────────────────────────────────┐│                                                                                                                                                                             
││         Total Time: 0.298s        ││                                                                                                                                                                             
│└───────────────────────────────────┘│                                                                                                                                                                             
└─────────────────────────────────────┘                                                                                                                                                                             
┌───────────────────────────┐                                                                                                                                                                                       
│      EXPLAIN_ANALYZE      │                                                                                                                                                                                       
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │                                                                                                                                                                                       
│             0             │                                                                                                                                                                                       
│          (0.00s)          │                                                                                                                                                                                       
└─────────────_─────────────┘                                                                                                                                                                                       
┌─────────────┴─────────────┐                                                                                                                                                                                       
│         PROJECTION        │                                                                                                                                                                                       
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │                                                                                                                                                                                       
│ jaccard(column1, column2) │                                                                                                                                                                                       
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │                                                                                                                                                                                       
│          2000000          │                                                                                                                                                                                       
│          (0.69s)          │                                                                                                                                                                                       
└─────────────_─────────────┘                                                                                                                                                                                       
┌─────────────┴─────────────┐                                                                                                                                                                                       
│READ_CSV_AUTO (MULTI-T...  │                                                                                                                                                                                       
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │                                                                                                                                                                                       
│          column1          │
│          column2          │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│        EC: 4916324        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (5.51s)          │
└───────────────────────────┘      
```