@carlopi 2023-11-03
We can likely get another order of magnitude faster using `std::bitset<MAX_SIZE>` instead of `std::array<char, MAX_SIZE>`.