@ucasfl 2023-11-03
> We can likely get another order of magnitude faster using `std::bitset<MAX_SIZE>` instead of `std::array<char, MAX_SIZE>`.

Not see improvement but degrade with `bitset`:

```
D explain analyze select jaccard(column1, column2) from 'data.csv';

┌─────────────────────────────┐
│┌───────────────────────────┐│
│└───────────────────────────┘│
└─────────────────────────────┘
┌─────────────────────────────────────┐
│┌───────────────────────────────────┐│
││    Query Profiling Information    ││
│└───────────────────────────────────┘│
└─────────────────────────────────────┘
explain analyze select jaccard(column1, column2) from 'data.csv';
┌─────────────────────────────────────┐
│┌───────────────────────────────────┐│
││         Total Time: 0.262s        ││
│└───────────────────────────────────┘│
└─────────────────────────────────────┘
┌───────────────────────────┐
│      EXPLAIN_ANALYZE      │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│             0             │
│          (0.00s)          │
└─────────────_─────────────┘                             
┌─────────────┴─────────────┐
│         PROJECTION        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│ jaccard(column1, column2) │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (2.47s)          │
└─────────────_─────────────┘                             
┌─────────────┴─────────────┐
│READ_CSV_AUTO (MULTI-T...  │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          column1          │
│          column2          │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│        EC: 4916324        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (5.04s)          │
└───────────────────────────┘     
```