@carlopi 2023-11-03
Can you share how to generate `data.csv`?