@ucasfl 2023-11-03
> Can you share how to generate `data.csv`?

https://clickhouse.com/docs/en/getting-started/example-datasets/star-schema

```
 ./dbgen -s 1000 -T s
```