@carlopi 2023-11-03
I was exagerating, but this (on my machine!!) is somehow faster (20-30%), and I would expect this to be generally faster / arguably more readable:
```c++
static inline void GetSet(const string_t &str, std::bitset<MAX_SIZE> & array_set) {
        array_set.reset();

        idx_t str_len = str.GetSize();
        auto s = str.GetData();

        for (idx_t pos = 0; pos < str_len; pos++) {
                array_set.set(static_cast<unsigned char>(s[pos]));
        }
}

static double JaccardSimilarity(const string_t &str, const string_t &txt) {
        if (str.GetSize() < 1 || txt.GetSize() < 1) {
                throw InvalidInputException("Jaccard Function: An argument too short!");
        }
        std::bitset<MAX_SIZE> m_str, m_txt;

        GetSet(str, m_str);
        GetSet(txt, m_txt);

        std::bitset<MAX_SIZE> m_and = m_str;
        m_and &= m_txt;

        std::bitset<MAX_SIZE> m_or = m_str;
        m_or |= m_txt;

        idx_t size_intersect = m_and.count();
        idx_t size_union = m_or.count();

        return static_cast<double>(size_intersect) / static_cast<double>(size_union);
}
```