@ucasfl 2023-11-03
I was wrong, it has better performance:
```sql
explain analyze select jaccard(column1, column2) from 'data.csv';
┌─────────────────────────────────────┐
│┌───────────────────────────────────┐│
││         Total Time: 0.196s        ││
│└───────────────────────────────────┘│
└─────────────────────────────────────┘
┌───────────────────────────┐
│      EXPLAIN_ANALYZE      │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│             0             │
│          (0.00s)          │
└─────────────_─────────────┘                             
┌─────────────┴─────────────┐
│         PROJECTION        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│ jaccard(column1, column2) │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (0.25s)          │
└─────────────_─────────────┘                             
┌─────────────┴─────────────┐
│READ_CSV_AUTO (MULTI-T...  │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          column1          │
│          column2          │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│        EC: 4916324        │
│   ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─   │
│          2000000          │
│          (5.22s)          │
└───────────────────────────┘  
```